+++ react-ui/src/components/Search_Results/Search_Results.css
