+++ react-ui/src/components/Search_Results/Search_Results.js
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import Button from '@material-ui/core/Button';
 import { Link } from 'react-router-dom'
-import './Restaurant-info.css';
+import './Search_Results.css';
 
-export const RestaurantInfo = (props) => {
+export const SearchResults = (props) => {
     const {id} = props.match.params
     const [details, setDetails] = useState({})
